% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_null}
\alias{clean_null}
\title{Clean rds files with NULL values}
\usage{
clean_null(files)
}
\arguments{
\item{files}{\code{(chr)} of files to clean NULL saves}
}
\value{
\code{(chr)} of valid files after cleaning
}
\description{
Clean rds files with NULL values
}
