% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_fn_utils.R
\name{create_accessors}
\alias{create_accessors}
\title{Create accessor functions in the namespace}
\usage{
create_accessors(
  dep_dir = "data",
  deps = NULL,
  dep_update = dep_update_fn(fn = dep_update_dropbox),
  update_all = TRUE
)
}
\arguments{
\item{dep_dir}{\code{(chr)} of the directory in which dependencies to be accessed with functions is stored. \emph{Note:} Either \code{dep_dir} or \code{deps} must be supplied.}

\item{deps}{\code{(chr)} of full filepaths for all dependencies to be accessed with functions.}

\item{dep_update}{\code{(function)} dependency update from remote location function}

\item{update_all}{\code{(lgl)} update all dependencies immediately.}
}
\value{
The accessor functions will be available in the package namespace if a \code{golem} package, otherwise accessor functions will be returned in a list
}
\description{
Create accessor functions in the namespace
}
