% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_fns.R
\name{db_auth}
\alias{db_auth}
\title{Authorize Dropbox}
\usage{
db_auth(
  db_token = file.path(pkgload::pkg_path(), "inst", "vault", "db_token.rds")
)
}
\arguments{
\item{db_token}{\code{(character)} path to Dropbox token saved as RDS}
}
\description{
Authorize Dropbox
}
