% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_update_fns.R
\name{dep_update_dropbox}
\alias{dep_update_dropbox}
\title{Update a dependency saved in Dropbox (if it's older)}
\usage{
dep_update_dropbox(
  dep_filepath = file.path("data", paste0("deps_", pkgload::pkg_name(), ".rds")),
  dest_dir = "data"
)
}
\arguments{
\item{dep_filepath}{\code{(chr)} of the filepath}

\item{dest_local_dir}{\code{(dest_dir)} directory where the downloaded file will be saved.}

\item{dest_remote_dir}{\code{(dest_dir)} directory on the remote server where \code{dep_filepath} is saved. The location will be transformed as follows: \code{\link[base]{file.path}(dest_remote_dir, \link[base]{basename}(dep_filepath)}}
}
\value{
dep_filepath contents
}
\description{
Update a dependency saved in Dropbox (if it's older)
}
