% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_update_fns.R
\name{dep_update_fn}
\alias{dep_update_fn}
\title{Construct a dependency update function}
\usage{
dep_update_fn(args, body, fn)
}
\arguments{
\item{args}{\code{(pairlist)} See \link[rlang]{pairlist2}}

\item{body}{\code{(expression)} See \link[rlang]{new_function}}

\item{fn}{\code{(function)} Use to pass the default dependency update function}
}
\value{
\code{(function)}
}
\description{
Construct a dependency update function
}
\seealso{
rlang::new_function
}
