% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_update_fns.R
\name{update_dropbox}
\alias{update_dropbox}
\title{Update a single file from dropbox}
\usage{
update_dropbox(local_filepath, remote_fp = NULL)
}
\arguments{
\item{local_filepath}{\code{(chr)} path to local file}

\item{remote_fp}{\code{(chr)} path to remote file (if different)}

\item{do_update}{\code{(lgl)} Whether to do this update. If files are all updated on app load see (\code{create_accessors} argument \code{update_all}), this function can be muted.}
}
\value{
Nothing
}
\description{
Update a single file from dropbox
}
